Enum minivm::backend::bytecode::Instruction

pub enum Instruction {
    Nop,
    IPush,
    IStore,
    ILoad,
    IAdd,
    Print,
}

Defines the available byte code instructions.

Inspired by some others:

Java Byte Code

Microsoft Common Intermediate Language

Lua 5.1 VM Instructions

Smalltalk-80 Blue Book

Variants
Nop

Perform no operation. Stack (before -> after): [no change] Other bytes (count: operand labels): -

IPush

Pushes eight bytes onto the stack as an integer value. Stack (before -> after): [] -> [value] Other bytes (count: operand labels): 8: byte1, byte2, byte3, byte4, byte5, byte6, byte7, byte8

IStore

Store integer value into variable #index. Stack (before -> after): [value] -> [] Other bytes (count: operand labels): 1: index

ILoad

Load an integer value from a local variable #index. Stack (before -> after): [] -> [value] Other bytes (count: operand labels): 1: index

IAdd

Add two integers. Stack (before -> after): [value1, value2] -> [result] Other bytes (count: operand labels): -

Print

Print the value on top of the stack. /// Stack (before -> after): [value] -> [] Other bytes (count: operand labels): -