Enum minivm::backend::bytecode::List

pub enum List {
    Cons(Instruction, Box<List>),
    Nil,
}

Variants
Cons(Instruction, Box<List>)

Tuple struct that wraps an element and a pointer to the next node.

Nil

A node that signifies the end of the linked list.

Methods

impl List

pub fn new() -> List

pub fn prepend(self, elem: Instruction) -> List

pub fn append(self, elem: Instruction) -> List

pub fn len(&self) -> u32

pub fn stringify(&self) -> String