Struct minivm::frontend::character_helper::CharacterHelper

pub struct CharacterHelper {}

Provides several helper functions to recognize characters.

Methods

impl CharacterHelper

pub fn is_alphabetic(ch: char) -> bool

Determines if a given character is alphabetic.

pub fn is_numeric(ch: char) -> bool

Determines if a given character is numeric.

pub fn is_alphanumeric(ch: char) -> bool

Determines if a given character is alphabetic or numeric.

pub fn is_double_quote(ch: char) -> bool

Determines if a given character is a double quote.

pub fn is_single_quote(ch: char) -> bool

Determines if a given character is a single quote.

pub fn is_operator(ch: char) -> bool

Determines if a given character is an operator.

pub fn is_white_space(ch: char) -> bool

Determines if a given character is a white space.

pub fn is_new_line(ch: char) -> bool

Determines if a given character is a new line.